Update codecs. Correct data types.

+++ common/src/main/scala/io/apex/f1/data/LobbyInfoData.scala
@@ -3,7 +3,7 @@ package io.apex.f1.data
 import io.apex.f1.enums.*
 
 case class LobbyInfoData(
-    aiControlled: Short,
+    aiControlled: Int,
     teamId: Team,
     nationality: Nationality,
     name: String,
